+++ example-lang/zh_TW/passwords.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Password Reset Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are the default lines which match reasons
+    | that are given by the password broker for a password update attempt
+    | outcome such as failure due to an invalid password / reset token.
+    |
+    */
+
+    'reset' => '您的密碼已重設。',
+    'sent' => '我們已將密碼重設連結寄給您。',
+    'throttled' => '請稍候再試。',
+    'token' => '此密碼重設憑證無效。',
+    'user' => '找不到使用該電子郵件地址的用戶。',
+
+];